clean up codes for compute instance and bms
as the schemas of compute and bms are similar, we can merge some
methdods.
